@@ -1,6 +1,6 @@
 <?php
 
-$EM_CONF[$_EXTKEY] = array(
+$EM_CONF[$_EXTKEY] = [
     'title' => 'Extbase Filter',
     'description' => 'Ability to use property filters.',
     'category' => 'system',
@@ -20,12 +20,12 @@ $EM_CONF[$_EXTKEY] = array(
     'lockType' => '',
     'author_company' => '',
     'version' => '1.1.1',
-    'constraints' => array(
-        'depends' => array(
-            'typo3' => '6.2.0-0.0.0',
-            'php' => '5.3.0-0.0.0',
-        ),
-        'conflicts' => array(),
-        'suggests' => array(),
-    ),
-);
+    'constraints' => [
+        'depends' => [
+            'typo3' => '6.2.0-8.7.99',
+            'php' => '5.5.0-0.0.0',
+        ],
+        'conflicts' => [],
+        'suggests' => [],
+    ],
+];
